\name{SCArray-classes}
\alias{SCArray-classes}
\alias{SCArrayFileClass}
\alias{SCArrayFileClass-class}
\alias{SC_GDSArray}
\alias{SC_GDSArray-class}
\alias{SC_GDSMatrix}
\alias{SC_GDSMatrix-class}

\docType{class}
\title{Class list defined in SCArray}
\description{
    \code{SCArrayFileClass} is a class directly inheriting from
\code{gds.class}. \code{SC_GDSArray} is a DelayedArray with a
\code{SCArraySeed}. \code{SC_GDSMatrix} is 2-dim \code{SC_GDSArray}.

    The package combines dense/sparse matrices stored in GDS files and the
Bioconductor infrastructure framework to provide out-of-memory data storage
and manipulation using the R programming language.
}
\author{
    Xiuwen Zheng \email{xiuwen.zheng@abbvie.com}
}

\keyword{GDS}
\keyword{SingleCell}
