\name{scClose}
\alias{scClose}
\title{Close the Single-cell GDS File}
\description{
    Closes a single-cell GDS file which is open.
}
\usage{
scClose(gdsfile)
}
\arguments{
    \item{gdsfile}{a single-cell GDS object with class \code{SCArrayFileClass}}
}
\value{
    None.
}

\author{Xiuwen Zheng}
\seealso{
    \code{\link{seqOpen}}
}

\keyword{GDS}
\keyword{SingleCell}
