\name{SCArray-package}
\alias{SCArray-package}
\alias{SCArray}
\docType{package}
\title{Large-scale single-cell RNA-seq data manipulation and analysis with
efficient algorithms}
\description{
    Large-scale single-cell RNA-seq data manipulation and statistical analysis
with scalable implementation of generalized mixed models and principal component
analysis. The package integrates sparse matrices in Genomic Data Structure (GDS)
files and the Bioconductor infrastructure framework to provide out-of-memory
data storage and manipulation in R.
}
\details{
    \tabular{ll}{
        Package: \tab SCArray \cr
        Type: \tab Package \cr
        License: \tab GPL version 3 \cr
    }
}
\references{
}
\author{
    Xiuwen Zheng \email{xiuwen.zheng@abbvie.com}
}

\examples{
}

\keyword{GDS}
\keyword{SingleCell}
